feat(storybook): scope live agent session to AgentSessionView-01 only
Splits AgentSessionProvider into two decorators: the default
AgentSessionProvider goes back to the shared sandbox token server (dummy
session, no local setup required) for every story except AgentSessionView-01,
which now uses a new LiveAgentSessionProvider that hits the local
/api/agents-ui/token dev route and dispatches AGENT_NAME against a real
LiveKit project.
