+++ docs/storybook/.env.example
@@ -1,3 +1,9 @@
+# For agents-ui, create a sandbox token server on cloud.livekit.io
+# Used by most stories, since it works without a local LiveKit project.
+VITE_PUBLIC_LK_SANDBOX_TOKEN_SERVER_ID=
+
+# Used only by the AgentSessionView-01 story, which connects to a real local
+# LiveKit project so it can be exercised against a real agent.
 LIVEKIT_URL=
 LIVEKIT_API_KEY=
 LIVEKIT_API_SECRET=
