+++ docs/storybook/.storybook/env.d.ts
@@ -1,5 +1,6 @@
 /// <reference types="vite/client" />
 interface ImportMetaEnv {
+  readonly VITE_PUBLIC_LK_SANDBOX_TOKEN_SERVER_ID: string;
   // more env variables...
 }
