+++ docs/storybook/stories/agents-ui/AgentSessionView-01.stories.tsx
@@ -1,13 +1,12 @@
 import React from 'react';
 import { StoryObj } from '@storybook/react-vite';
 import { useTheme } from 'next-themes';
-import { AgentSessionProvider } from '../../.storybook/lk-decorators/AgentSessionProvider';
+import { LiveAgentSessionProvider } from '../../.storybook/lk-decorators/LiveAgentSessionProvider';
 import { AgentSessionView_01, AgentSessionView_01Props } from '@livekit/agents-ui';
-
 
 export default {
   component: AgentSessionView_01,
-  decorators: [AgentSessionProvider],
+  decorators: [LiveAgentSessionProvider],
   render: (args: AgentSessionView_01Props) => {
     const { resolvedTheme = 'dark' } = useTheme();
     return <AgentSessionView_01 themeMode={resolvedTheme as 'dark' | 'light'} {...args} />;
